fix: build verification — 8 compilation/logic issues + session 4 hidden audit fixes
Session 4 (hidden audit):
- H-AUDIT-1/2: Per-session synchronized locks in SessionService
- M-AUDIT-3: MatchmakingService queue atomicity
- M-AUDIT-4: LobbyWebSocketHandler ready atomicity
- M-AUDIT-5: ChatWebSocketHandler Long truncation
- M-AUDIT-6: ChatController auth checks session participation
- M-AUDIT-7: ReplayController 14MB upload limit
- H-AUDIT-9: Control group recall actually selects units
- M-AUDIT-10: MatchmakingService uses 'test_map' not 'default'

Session 5 (build verification):
- BUILD-1: GameWebSocketHandler 4 record field accesses missing ()
- BUILD-2: Long reference equality != → .equals()
- BUILD-3: ChatControllerTest missing SessionService mock
- BUILD-4: ChatControllerTest missing Authentication param
- BUILD-5: StatsRegistry removed Guava @VisibleForTesting
- BUILD-6: SessionService removed duplicate @PostConstruct
- BUILD-7: ResearchRegistry removed e.printStackTrace()
- BUILD-8: MainMenuScene + DataOverride unused imports

Updated: TODO.md (100 issues, 95 fixed), ProjectProgress.md, AGENT.md

+++ aow2-server/src/main/java/com/aow2/server/websocket/ChatWebSocketHandler.java
@@ -217,7 +217,7 @@ public class ChatWebSocketHandler extends TextWebSocketHandler {
         long timestamp = chatMessage.getTimestamp().toEpochMilli();
         Map<String, Object> broadcastMsg = Map.of(
                 "type", "chat",
-                "playerId", playerId.intValue(),
+                "playerId", playerId,
                 "message", message,
                 "timestamp", timestamp
         );
